Fix error happening on windows

+++ src/timberman/Log.py
@@ -14,12 +14,18 @@ class Log:
     def __init__(self, screen_handle: MSSBase, position: int):
         self.__position = position
         self.__screen_handle = screen_handle
-        self.__left_branch_bounding_box = {"top": SystemProperties.screen_height - SystemProperties.tree_offset - position * Log.height + Log.skip_top_pixels,
-                                           "left": SystemProperties.screen_width / 2 - Log.width / 2 - Log.branch_width, "width": Log.branch_width - Log.offset_from_log,
-                                           "height": Log.height - Log.skip_top_pixels}
-        self.__right_branch_bounding_box = {"top": SystemProperties.screen_height - SystemProperties.tree_offset - position * Log.height + Log.skip_top_pixels,
-                                            "left": SystemProperties.screen_width / 2 + Log.width / 2 + Log.offset_from_log, "width": Log.branch_width - Log.offset_from_log,
-                                            "height": Log.height - Log.skip_top_pixels}
+        self.__left_branch_bounding_box = {
+            "top": int(SystemProperties.screen_height - SystemProperties.tree_offset - position * Log.height + Log.skip_top_pixels),
+            "left": int(SystemProperties.screen_width / 2 - Log.width / 2 - Log.branch_width),
+            "width": int(Log.branch_width - Log.offset_from_log),
+            "height": int(Log.height - Log.skip_top_pixels)
+        }
+        self.__right_branch_bounding_box = {
+            "top": int(SystemProperties.screen_height - SystemProperties.tree_offset - position * Log.height + Log.skip_top_pixels),
+            "left": int(SystemProperties.screen_width / 2 + Log.width / 2 + Log.offset_from_log),
+            "width": int(Log.branch_width - Log.offset_from_log),
+            "height": int(Log.height - Log.skip_top_pixels)
+        }
         self.take_screenshot()
 
     def take_screenshot(self):
